Core/Mail: - Fix mailing items - Fix updating mail messages (mark as read, etc.) - Fix Player::SaveGoldToDB

+++ src/server/game/Mails/Mail.cpp
@@ -213,8 +213,8 @@ void MailDraft::SendMailTo(SQLTransaction& trans, MailReceiver const& receiver, 
     stmt->setBool  (++index, !_items.empty());
     stmt->setUInt64(++index, uint64(expire_time));
     stmt->setUInt64(++index, uint64(deliver_time));
-    stmt->setUInt32(++index, _money);
-    stmt->setUInt32(++index, m_COD);
+    stmt->setUInt64(++index, _money);
+    stmt->setUInt64(++index, m_COD);
     stmt->setUInt8 (++index, uint8(checked));
     trans->Append(stmt);
